Generator plugin implemented.
