+++ vaadin-connect-demo/src/test/java/com/vaadin/connect/demo/VaadinConnectControllerIT.java
@@ -26,7 +26,6 @@ import java.util.Optional;
 import java.util.stream.Stream;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -272,8 +271,6 @@ public class VaadinConnectControllerIT {
   }
 
   @Test
-  @Ignore // Re-enable after move swagger-codegen to maven plugin module
-  @SuppressWarnings("squid:S1607")
   public void should_RequestFail_When_InvalidRoleUsedInRequest() {
     String methodName = "permitRoleAdmin";
 
@@ -284,8 +281,6 @@ public class VaadinConnectControllerIT {
   }
 
   @Test
-  @Ignore // Re-enable after move swagger-codegen to maven plugin module
-  @SuppressWarnings("squid:S1607")
   public void should_RequestFail_When_MethodCallProhibitedByClassAnnotation() {
     String methodName = "deniedByClass";
