+++ vaadin-connect-maven-plugin/src/main/resources/connect-client.default.js.template
