+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/application.properties.for.testing
