+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/expected-connect-client-custom.js
