+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/expected-first-class-multiple-tags.js
@@ -1,0 +1,26 @@
+import defaultClient from './connect-client.default';
+
+export class MyFirstJsClass {
+
+  /**
+   * Create a MyFirstJsClass instance.
+   * @param {ConnectClient=} client - an instance of ConnectClient
+   */
+  constructor(client = defaultClient) {
+    this._client = client;
+  }
+
+  /**
+   * Get all users
+   *
+   * @returns {array} Return list of users
+   */
+  getAllUsers() {
+    return this._client.call('GeneratorTestClass', 'getAllUsers');
+  }
+
+}
+
+const service = new MyFirstJsClass(defaultClient);
+
+export const getAllUsers = service.getAllUsers.bind(service);
