+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/expected-openapi-custom-application-properties.json
