+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/get-operation-openapi.json
