+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/multiple-tags-operation.json
