--- vaadin-connect/src/main/resources/META-INF/services/io.swagger.codegen.v3.CodegenConfig
@@ -1,1 +1,0 @@
-com.vaadin.connect.generator.VaadinConnectJsGenerator
